change interface of withAsyncReducer
simply change withAsyncReducer so it is a function inside a function.
this interface mirrors other HOCs (like connect) where the first call
to the function represents how you are configuring it and the second
call is what you are wrapping. In the case of HOCs, you wrap a
component. But in the case of HORs, you wrap the reducer