delete HORs withAsyncReducer and instead use helpers from @reduxjs/toolkit
withAsyncReducer was confusing to students, so delete it. in its place
there is now a /stateReducers/helpers folder with asyncCases,
asyncInitialState, and getInitStateFromStorage.

asyncCases represent the three cases typically required for asynchronous
actions. so asyncCases covers the same use-case as withAsyncReducer,
except it doesnt actually create a reducer--it just provides the three
case functions. These asycCases can then be passed to a createReducer
call.

createReducer comes from @reduxjs/toolkit as a helper function which
reduces the amount of code needed to write a reducer. it does this
primarily by removing the need to write a switch statement with literal
cases and instead just provide an object with properties for each case
whose value is a function.

asyncInitialState was abstracted into /stateReducers/helpers in order to
be re-usable for any reducer which uses asyncCases.

getInitialStateFromStorage was also abstracted into
/stateReducer/helpers. it reads a key from localStorage and hydrates
the inital reducer state with that value (if it finds one). this helper
is only being used in login, but abstracted it into helpers in case
someone wants to easily re-use it.