update action types to use createAction so they are action creators
createAction comes from @reduxjs/toolkit. it helps reduces the amount of
code that must be written by creating a callable function. the callable
function expects the action payload to be passed. This reduces the
amount of code in the async action creators because you do not have to
create the action object to put the type property into it. the action
object is created with the type property already in it when you call a
function created by createAction.

action types are now following the new recommended style guide https://redux.js.org/style-guide/style-guide/#write-action-types-as-domaineventname