refactor /redux for simplicity
significantly re-arrange files in the redux folder to reduce the number
of files and sub-folders. there are now just single files in /redux for
auth.js, users.js, messages.js, likes.js. inside of each should contain
all the action creator functions for that part of the API as well as the
reducers for each of those async action creators.

helpers which were in /stateReducers/helpers were moved into /helpers.
so now /helpers contains hepler for both action creators and reducers

delete configureStore.js and instead put all the setup of the store in
/redux/index.js